scripts/pyproject.toml
[build-system]
requires = ["flit_core==3.9.0"]
build-backend = "flit_core.buildapi"

[project]
name = "boilercv_tools"
version = "0.0.0"
description = "Tools."
dependencies = [
    # ? Core script tools
    "copier==9.2.0",
    "cyclopts==2.9.3",
    "dulwich==0.22.1",
    "ipython==8.26.0",
    "pipx==1.6.0",
    "pytest-env==1.1.3",
    "python-dotenv==1.0.1",
    "toml==0.10.2 ; python_version < '3.11'",
]
[project.scripts]
"boilercv_tools" = "boilercv_tools.__main__:main"

[tool.fawltydeps]
deps = ["scripts/pyproject.toml"]
code = ["scripts/boilercv_tools"]
ignore_undeclared = [
    "boilercore",
    "boilercv_docs",
    "boilercv_pipeline",
    "boilercv_tests",
    "boilercv_tools",
    "boilercv",
]
ignore_unused = [
    # ? Core script tools
    "copier", # For template sync
    "pipx",   # For `pipx run` of scripts with dependencies isolated by PEP 723
    # ? Other deps
    "ipython", # Used for local experimentation
]
